Production stabilization: startup fixes, trading correctness, silent failure removal, security hardening
C1/C2/C3 — Fix undefined variable in signal_agent.py, add missing imports in main.py, fix strategy constructor mismatch
B1/B2/B6 — None-safe risk checks, settings fallbacks, deterministic stop-loss evaluation
C5/C6 — Docker scripts copy fix, compose volume mount fix
C8 — Arbitrary market outcome support in execution agent
B5 — Ensure attribution archetype always persisted
Float precision — Migrate safety.daily_pnl from Float to Numeric
Phase 3 — Replace silent except:pass with structured logging in events.py and event_bridge.py
Phase 4 — Register adaptive_meta strategy (was missing from __init__.py)
Phase 5 — Resilient lifespan startup, 8 integration tests added (14 passing)
Phase 6 — Debug endpoint guard for production, safe CORS, gitignored .env

+++ backend/app/models/safety.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import String, Boolean, DateTime, Integer, Float
+from sqlalchemy import String, Boolean, DateTime, Integer, Numeric
 from sqlalchemy import JSON
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
@@ -18,7 +18,7 @@ class SafetyState(Base):
     circuit_breaker_active: Mapped[bool] = mapped_column(Boolean, default=False)
     circuit_breaker_reason: Mapped[str | None] = mapped_column(String(64), nullable=True)
     quarantined_strategies: Mapped[list | None] = mapped_column(JSON, nullable=True, default=list)
-    daily_pnl: Mapped[float] = mapped_column(Float, default=0.0)
+    daily_pnl: Mapped[float] = mapped_column(Numeric(24, 8), default=0.0)
     checks_passed: Mapped[int] = mapped_column(Integer, default=0)
     checks_failed: Mapped[int] = mapped_column(Integer, default=0)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
